requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-groq
requests
markdown
pygments